Move the impl of CoordinatorGroupCommitter
to DefaultGroupCommitKeyManipulator to re-use it
